#693 intrinsic_swap_endian_16 was demanding a 32-bit argument instead of 16

+++ include/_DEVELOPMENT/clang/intrinsic.h
@@ -85,7 +85,7 @@ extern void *intrinsic_return_de(void);
 
 
 
-extern unsigned int intrinsic_swap_endian_16(unsigned long n);
+extern unsigned int intrinsic_swap_endian_16(unsigned int n);
 
 
 extern unsigned long intrinsic_swap_endian_32(unsigned long n);
